+++ OwnTracks/OwnTracks/FriendsTVC.m
@@ -18,8 +18,6 @@
 #import "FriendAnnotationV.h"
 #import "OwnTracking.h"
 #import <CocoaLumberjack/CocoaLumberjack.h>
-#import <Fabric/Fabric.h>
-#import <Crashlytics/Crashlytics.h>
 
 @interface FriendsTVC ()
 @property (strong, nonatomic) UIAlertView *alertView;
@@ -198,8 +196,6 @@ forRowAtIndexPath:(NSIndexPath *)indexPath {
         NSError *error = nil;
         if (![context save:&error]) {
             DDLogError(@"Unresolved error %@, %@", error, [error userInfo]);
-            [[Crashlytics sharedInstance] setObjectValue:@"deleteFriend" forKey:@"CrashType"];
-            [[Crashlytics sharedInstance] crash];
         }
     }
 }
@@ -234,8 +230,6 @@ forRowAtIndexPath:(NSIndexPath *)indexPath {
     NSError *error = nil;
     if (![self.fetchedResultsController performFetch:&error]) {
         DDLogError(@"Unresolved error %@, %@", error, [error userInfo]);
-        [[Crashlytics sharedInstance] setObjectValue:@"fetchFriends" forKey:@"CrashType"];
-        [[Crashlytics sharedInstance] crash];
     }
     
     return _fetchedResultsController;